Added gitlab_ci_pipeline_environment_deployment_count metric
